third_party/blink/web_tests/external/wpt/css/compositing/canvas-composite-modes-expected.txt

This is a testharness.js-based test.
PASS globalCompositeOperation clear
PASS globalCompositeOperation copy
FAIL globalCompositeOperation destination assert_equals: composite operation expected "destination" but got "source-over"
PASS globalCompositeOperation source-over
PASS globalCompositeOperation destination-over
PASS globalCompositeOperation source-in
PASS globalCompositeOperation destination-in
PASS globalCompositeOperation source-out
PASS globalCompositeOperation destination-out
PASS globalCompositeOperation source-atop
PASS globalCompositeOperation destination-atop
PASS globalCompositeOperation xor
FAIL globalCompositeOperation lighter assert_approx_equals: alpha: g=0 a=47 lighter g=0 a=255 expected 302 +/- 1.01 but got 255
FAIL globalCompositeOperation plus-darker assert_equals: composite operation expected "plus-darker" but got "source-over"
FAIL globalCompositeOperation plus-lighter assert_equals: composite operation expected "plus-lighter" but got "source-over"
Harness: the test ran to completion.
